Switch to anyhow+thiserror
We need special handling for very few errors (i.e not logged in), but in
general we need a mechanism to capture any error and return up, possibly
adding context for ease of debugging.

This commit switches to anyhow::Result (so anything can be returned).
Also, a few specific types are added via thiserror (DeepCtlError).

Currently only not-logged-in is "properly" handled.

Related to #36