Add deploy create shortcut
I'm personally not in favor of this. Now the user might pass the -m/-t
flags from the deploy subcommand when issuing other commands, and they
won't be honored. Also there is no good way to put documentation that
one is the same as the other.

All this to save 6 characters, it's not worth it. Also deploy is not
a command that you're supposed to run a lot. Infer -- yes, but deploy
... meh.